fix: decode RTAB-Map pose blobs
